requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
databases==0.9.0
ecdsa==0.19.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0

